Search and suggesting working.
Autocomplete works, but I have a hack to return only the suggestions. The query is
hard to test, or at least following the links is, because it takes me to the
real site, and not the one under development.

TODO: Fix the css to make it look nice.
TODO: Fix the hack to return every second term (because frequency is between).
TODO: Fix the terms query not to need the field on the query