Fixing which search engine to use.
The way I was switching search engines between dev and prod was fine,
but I forgot to allow switch of Solr cores. Fixed that. I have a slightly
new technique, but I only did it for the core so far. It could be applied
to the search engine address, too.

I also thought I had fixed the search button to be the default, but it
needs more testing.