Use Bootstrap.
Mostly working with Bootstrap. There was a weird positioning thing when
you use floating navbars. It was overlapping the container. I had to set
a margin in the body, documented in stylesheet.scss.

TODO: the markdown files don't seem to render right when I have a nested
render like I do now.

TODO: Could have made the choice of layouts parametrized by whether it's
the home page or not.