Stacking kramdown and erb.
This version has the erb filter first. It means you have to enable
kramdown processing inside HTML blocks, which is not the default. I found
you can put an attribute "markdown=1" in the HTML block level element
you want to parsing to be in. You could also do
`filter :kramdown, parse_block_html: true` in the Rules, but there were
warnings in the doc that globally enabling might lead to surprises.

TODO: Figure out why the way everyone else does it doesn't work for me,
particularly where it looks like escaping the "<" and ">" in the ERB in the
Markdown file didn't seem to work.