Much better.
It all magically works, since I put the different formatting in different
layouts, and figured out how to do that. The compile rule calls the layout.
The layout rule is chosen by matching to the symbol given as a argument
to the format method.

What's still not clear to me is why it was broken before. I suspect I must
have had some back html. Or it was the way kramdown was processing things.
I think I found what might be a bug in kramdown. When a list was at the end
of a document, it seemed to eat the first </div> of the layout.